Fix Jenkins errors.
