+++ cdm-test/src/test/java/ucar/nc2/grid/TestGridCompareProblem.java
@@ -7,7 +7,7 @@ import ucar.unidata.util.test.TestDir;
 import static ucar.nc2.grid.TestReadGridCompare.compareGridCoverage;
 import static ucar.nc2.grid.TestReadGridCompare.compareGridDataset;
 
-@Ignore("not fixed yet")
+// @Ignore("not fixed yet")
 public class TestGridCompareProblem {
 
   @Test
@@ -27,6 +27,13 @@ public class TestGridCompareProblem {
   @Test
   public void testProblem2() throws Exception {
     String filename = "/media/snake/Elements/data/grib/idd/namPolar90/NamPolar90.ncx4";
+    compareGridDataset(filename);
+    compareGridCoverage(filename);
+  }
+
+  @Test
+  public void testProblem4() throws Exception {
+    String filename = TestDir.cdmUnitTestDir + "conventions/ifps/HUNGrids.netcdf";
     compareGridDataset(filename);
     compareGridCoverage(filename);
   }
