+++ cdm-test/src/test/java/ucar/nc2/ncml/TestAggNested.java
@@ -5,6 +5,8 @@
 package ucar.nc2.ncml;
 
 import static com.google.common.truth.Truth.assertThat;
+
+import org.junit.Ignore;
 import org.junit.Test;
 import org.junit.experimental.categories.Category;
 import org.junit.runner.RunWith;
@@ -22,6 +24,7 @@ import java.io.IOException;
 import java.lang.invoke.MethodHandles;
 
 /** Test opening nested NcML with and without use of NetcdfDataset.initNetcdfFileCache. */
+@Ignore("Nested aggreations are not working")
 @Category(NeedsCdmUnitTest.class)
 @RunWith(JUnit4.class)
 public class TestAggNested {
@@ -53,6 +56,17 @@ public class TestAggNested {
       cache.showCache();
     } finally {
       NetcdfDatasets.shutdown();
+    }
+  }
+
+  @Test
+  public void TestProblem() throws IOException {
+    String filename = TestDir.cdmUnitTestDir + "ncml/nestedAgg/test.ncml";
+    System.out.printf("TestAggNested %s%n", filename);
+
+    try (NetcdfDataset ncd = NetcdfDatasets.openDataset(filename)) {
+      Variable time = ncd.findVariable("time");
+      assertThat(time.getSize()).isEqualTo(19723);
     }
   }
