+++ cdm/core/src/test/java/ucar/nc2/internal/dataset/TestDatasetClassifier.java
@@ -303,56 +303,4 @@ public class TestDatasetClassifier {
     }
   }
 
-  @Test
-  @Category(NeedsExternalResource.class)
-  public void testNamPolar() throws IOException {
-    String filename = "/media/snake/Elements/data/grib/idd/namPolar90/NAM_Polar_90km_20131203_0000.grib2";
-    try (FeatureDatasetCoverage covDataset = CoverageDatasetFactory.open(filename)) {
-      for (CoverageCollection cc : covDataset.getCoverageCollections()) {
-        if (cc.getName().endsWith("MRUTP")) {
-          assertThat(cc.getCoverageType()).isEqualTo(FeatureType.GRID);
-        }
-      }
-    }
-
-    try (NetcdfDataset ds = NetcdfDatasets.openDataset(filename)) {
-      Formatter errlog = new Formatter();
-      Optional<GridDatasetImpl> grido = GridDatasetImpl.create(ds, errlog);
-      assertThat(grido.isPresent()).isTrue();
-      GridDatasetImpl gridDataset = grido.get();
-      if (!Iterables.isEmpty(gridDataset.getGrids())) {
-        DatasetClassifier dclassifier = new DatasetClassifier(ds, errlog);
-        DatasetClassifier.CoordSysClassifier classifier =
-            dclassifier.getCoordinateSystemsUsed().stream().findFirst().orElse(null);
-        assertThat(classifier.getFeatureType()).isEqualTo(FeatureType.GRID);
-      }
-    }
-  }
-
-  @Test
-  @Category(NeedsExternalResource.class)
-  public void testNamPolarCollection() throws IOException {
-    String filename = "/media/snake/Elements/data/grib/idd/namPolar90/NamPolar90.ncx4";
-    try (FeatureDatasetCoverage covDataset = CoverageDatasetFactory.open(filename)) {
-      for (CoverageCollection cc : covDataset.getCoverageCollections()) {
-        if (cc.getName().endsWith("MRUTP")) {
-          assertThat(cc.getCoverageType()).isEqualTo(FeatureType.GRID);
-        }
-      }
-    }
-
-    try (NetcdfDataset ds = NetcdfDatasets.openDataset(filename)) {
-      Formatter errlog = new Formatter();
-      Optional<GridDatasetImpl> grido = GridDatasetImpl.create(ds, errlog);
-      assertThat(grido.isPresent()).isTrue();
-      GridDatasetImpl gridDataset = grido.get();
-      if (!Iterables.isEmpty(gridDataset.getGrids())) {
-        DatasetClassifier dclassifier = new DatasetClassifier(ds, errlog);
-        DatasetClassifier.CoordSysClassifier classifier =
-            dclassifier.getCoordinateSystemsUsed().stream().findFirst().orElse(null);
-        assertThat(classifier.getFeatureType()).isEqualTo(FeatureType.GRID);
-      }
-    }
-  }
-
 }
